feat: Add useStorage composable and implement it in HomeView
This commit adds a new file `useStorage.js` which contains the `useStorage` function. The function allows reading and writing values to localStorage based on a provided key. It also includes a watch effect to update the stored value whenever the value changes.

In `HomeView.vue`, the `useFlash` import is replaced with the newly added `useStorage` import. Two variables, `food` and `age`, are created using the `useStorage` function with default values of 'salad' and 0 respectively. These variables are then used in the template for input bindings.

These changes enable persistent storage of user inputs for favorite food and age in the HomeView component.